can monitor example
This shows the ELM able to return CAN signals over Dbus
The next step needs for the Dbus and serial interaction to be handled by
separate classes/threads/processes with an intermediary message que in
between. Right now it is problematic to give the dbus object direct serial
interaction. The intermediary message que should funnel commands and allow
the publishing of responses.