Add interpretation to main elm dbus module
It ain't pretty...but the watcher and interpretation
routines are now in the main module for the elm dbus. The module now emits
at signals by default, and interprets can frames by default. Either
should be able to be switched using a dbus method.

Quite a bit of work still remains to be done for the stability of the module.