Simplify elm module serial backend
This removes the old modules inherited from the python-elm library, and
simplifies the serial connections used by the module to grant more control
by the module.

Series of commands needing to be sent in quick succession is now done
by the elm serial class instead of the dbus module. This may be a
consequence of the dbus module handling some commands out of order.
The serial class is able to ensure that the commands are sent in the
order that is desired, which is crucial when the ELM must listen to
the CAN.