ELM module startup reliability
Series of commands needing to be sent in quick succession is now done
by the elm serial class instead of the dbus module. This may be a
consequence of the dbus module handling some commands out of order.
The serial class is able to ensure that the commands are sent in the
order that is desired, which is crucial when the ELM must listen to
the CAN.